requirements.txt
websocket_client
numpy
TA-Lib